resolve type conflicts

+++ 0x17-doubly_linked_lists/7-insert_dnodeint.c
@@ -12,7 +12,7 @@ dlistint_t *insert_dnodeint_at_index(dlistint_t **h, unsigned int idx, int n)
 {
 	dlistint_t *temp;
 	dlistint_t *new;
-	int len, c;
+	size_t len, c;
 
 	temp = *h;
 	c = 0;
